Needed SET SINGLE_USER WITH ROLLBACK IMMEDIATE
The restore was failing because the DB is in use. It will throw an error like this 

Msg 3101, Level 16, State 1, Line 6
Exclusive access could not be obtained because the database is in use.
Msg 3013, Level 16, State 1, Line 6
RESTORE DATABASE is terminating abnormally.

We need this change to sort the problem.